gui: Inline the contents of x_tabs_create().
